Remove membership info
Unlike the CNCF and AI landscapes, we don't need to track membership in the linux foundation.
If there was an organization where we were trying to track membership, this would be different.
But we don't, so let's just remove it for now.

Testing done:
- without this change, we get the error

```
FATAL:  Membership category "CNCF Members" does not have corresponding category in landscape.yml
```

- with this change, we don't